Better distinguish between post and category
The constructor currently forgets the difference between a single post
and a category with only one post in it by wrapping any single post in
an array.  This causes the category to get the same generated URL as
the post itself, instead of using the configured bundle permalink
template.

To fix this, this patch teaches the constructor to remember whether it
received only a single post or an array containing a single post in
it.